Input: rohm_bu26154: filter out ADC noise
If the touch interface is enabled, XP and YP lines are Hi-Z so ADC
results are invalid. To filter out the invalid values, we have to check
the IRQB line to know if the screen is still touched after reading the
values, then discard them if the screen is no longer touched.

Signed-off-by: Ryosuke Saito <ryosuke.saito@sato-global.com>