Kernel: trace: Fix race condition reading saved tgids
Introduced ability to change saved cmdlines size (saved_cmdlines_size).
This resized saved command lines but missed resizing tgid mapping as well.

Another issue is that when the resize happens, it removes saved command
lines and reallocates new memory for it. This introduced a race
condition when reading the global savecmd as this can be freed in the
middle of accessing it causing a use after free access. Fix this by
implementing locking strategy.